Add option to start-paused
This option allows user to start omxplayer in a paused state instead of immediately playback the video.  The user can then use
standard keyboard or dbus commands to begin playback manually.

To use this function add the command-line parameter: --start-paused to
the omxplayer command.

Example:

omxplayer -p -o hdmi --start-paused test.mkv

This will start the omxplayer, load the buffer to begin playback and then pause.  To resume playback, press the key on the keyboard for resume (by default this is the p or space keys), or send a dbus command.